python_tests/requirements.txt
# General
pytest
allure-pytest
faker

# API
requests

# WebUI
webdriver_manager
splinter
selenium
